fixed images
I fixed the images/font issue. So the issue with the images on the index
were the resizable thing. I resized all the images to be the same size
so that when they are resized using vh, they will all have the same font
size. This way, we can still have the crisp sharp text of the image but
the resizabiltiy of the font.