Add smooth scrollTo, content to themizerindex.php
NOTE:

By downloading the files associated with this commit onto your personal
computer, you hereby agree to release me from being held liable for any
damages, dirty stares, and/or condemnation caused by any of (1)
giggling, (2) uncontrollable laughter, or (3) falling off your chair,
which in turn are caused by the files mentioned above and/or their
content. This is not a legally binding agreement, and is only for the
purpose(s) of parody. As a result, please do not attempt to treat this
as such.